backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.18
httpx==0.28.1
redis==5.2.1
celery[redis]==5.4.0
aiofiles==24.1.0
python-slugify==8.0.4
markdown==3.7
pygments==2.18.0
slowapi==0.1.9
